new code

+++ backend/controllers/goalControllers.js
@@ -7,7 +7,7 @@ const Goal = require('../models/goalModel')
 //@access Private
 const getGoals = asyncHandler (async(req,res)=>{
 
-    const goals = await Goal.find()
+    const goals = await Goal.find({user:req.user.id})
     
     res.status(200).json(goals)
 })
@@ -26,6 +26,7 @@ const setGoals = asyncHandler(async(req,res)=>{
 
    const goal = await Goal.create({
     text: req.body.text,
+    user:req.user.id,
    })
     res.status(200).json(goal)
 })
@@ -45,6 +46,22 @@ const updatetGoals = asyncHandler(async(req,res)=>{
         throw new Error('Goal not found')
     }
 
+
+    const user = await UserActivation.findById(req.user.id)
+
+//check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+
+    }
+
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+            res.status(401)
+            throw new Error('User not authorized')
+    }
+
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id,req.body,{new:true})
 
     res.status(200).json(updatedGoal)
@@ -57,11 +74,27 @@ const updatetGoals = asyncHandler(async(req,res)=>{
 //@access Private
 
 const deleteGoals = asyncHandler(async(req,res)=>{
+   
     const goal = await Goal.findById(req.params.id)
+
+
 
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    //check for user
+    if(!user){
+        res.status(401)
+        throw new Error('User not found')
+
+    }
+
+    //make sure the logged in user matches the goal user
+    if(goal.user.toString() !== user.id){
+            res.status(401)
+            throw new Error('User not authorized')
     }
 
     await goal.deleteOne()
